introduce Abstract post-commit hook receiver, allows to implement alternate web-hook payload computation introduced for cloudforge-plugin development
